+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -666,6 +666,16 @@ class GeckoEngine(
                 field = value
             }
 
+        override var cookieBannerHandlingDetectOnlyMode: Boolean = false
+            set(value) {
+                with(runtime.settings.contentBlocking) {
+                    if (this.cookieBannerDetectOnlyMode != value) {
+                        this.cookieBannerDetectOnlyMode = value
+                    }
+                }
+                field = value
+            }
+
         override var remoteDebuggingEnabled: Boolean
             get() = runtime.settings.remoteDebuggingEnabled
             set(value) { runtime.settings.remoteDebuggingEnabled = value }