+++ mobile/android/android-components/components/concept/engine/src/test/java/mozilla/components/concept/engine/SettingsTest.kt
@@ -168,6 +168,7 @@ class SettingsTest {
             enterpriseRootsEnabled = true,
             cookieBannerHandlingMode = EngineSession.CookieBannerHandlingMode.DISABLED,
             cookieBannerHandlingModePrivateBrowsing = EngineSession.CookieBannerHandlingMode.REJECT_ALL,
+            cookieBannerHandlingDetectOnlyMode = true,
         )
 
         assertFalse(defaultSettings.domStorageEnabled)
@@ -198,10 +199,10 @@ class SettingsTest {
         assertFalse(defaultSettings.fontInflationEnabled!!)
         assertEquals(2.0F, defaultSettings.fontSizeFactor)
         assertTrue(defaultSettings.forceUserScalableContent)
-        assertTrue(defaultSettings.loginAutofillEnabled)
         assertEquals(Color.BLUE, defaultSettings.clearColor)
         assertTrue(defaultSettings.enterpriseRootsEnabled)
         assertEquals(EngineSession.CookieBannerHandlingMode.DISABLED, defaultSettings.cookieBannerHandlingMode)
         assertEquals(EngineSession.CookieBannerHandlingMode.REJECT_ALL, defaultSettings.cookieBannerHandlingModePrivateBrowsing)
+        assertTrue(defaultSettings.cookieBannerHandlingDetectOnlyMode)
     }
 }